warnings

+++ Classes/Telemetry/BITSessionStateData.m
@@ -40,7 +40,7 @@
   [super encodeWithCoder:coder];
   [coder encodeObject:self.envelopeTypeName forKey:@"envelopeTypeName"];
   [coder encodeObject:self.dataTypeName forKey:@"dataTypeName"];
-  [coder encodeInt:self.state forKey:@"self.state"];
+  [coder encodeInt:(int)self.state forKey:@"self.state"];
 }
 
 @end
